requirements.txt
pydub
numpy
requests
ffmpeg-python
python-dotenv

